+++ lib/src/text.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/widgets.dart';
 
-import 'media.dart';
+import 'context.dart';
 
 extension WrapDirectionality on Widget {
   Widget wrapDirectionality(TextDirection direction) => Directionality(