+++ lib/wrap.dart
@@ -5,6 +5,5 @@ export 'src/decorate.dart';
 export 'src/handler.dart';
 export 'src/inherit.dart';
 export 'src/list.dart';
-export 'src/media.dart';
 export 'src/size.dart';
 export 'src/text.dart';