+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **Incompatible**: API about `ReContext` changed.
 - Use `builder` instead of `ReContext`, which is conciser.
+- Optimize tests and tidy code structure.
 
 ## 0.6.0